requirements.txt

# Django base
Django>=4.2

# DRF
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-coreapi==2.3.0
drf-yasg==1.21.8

# Django tools
django-filter==24.3

#Environ
django-environ==0.11.2